requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
numpy
faiss-cpu
datasets
requests
matplotlib
seaborn
pandas
